Add mongoDB to persist covid19 data
