+++ .env
@@ -1,2 +1,2 @@
-MONGOURI=mongodb+srv://dvd:dvd6900@cluster0.huu8r.mongodb.net/golangdb?retryWrites=true&w=majority
+MONGOURI=mongodb+srv://dvd:dvd6900@cluster0.titpd.mongodb.net/covid19IndiagoDB?retryWrites=true&w=majority
 PORT=3000
